Moved utils into target folders.
